Add SpectreGuard implementation
bit 9 in the pagetable now causes completed load instructions to not wake up dependents until they are at the head of the ROB. The term (W)rite-(B)ack-on-retire was used to signify this property, even though the value of the load is written back to the register file before it is retired. This allows secrets to be loaded into the processor speculatively, but does not allow its value to be used speculatively, which stops attacker from storing the value in side channels.